"Add select element with options and corresponding JavaScript functionality in index.html"
Explanation:
The code diff shows the addition of a select element with four options in the index.html file. A JavaScript function has also been added to handle the onchange event of the select element. When the first option is selected, the user is redirected to google.com. When the first option is selected, an alert box is displayed with a message and a quote.